calibrate

Function to estimate the resistivities and thicnesses of the layers.

calibrate(ves, opt.method = c("L-BFGS-B", "SA", "GA", "PSO", "DE"),
  obj.fn = c("rss", "mnad", "mxad", "log_rss", "log_mnad", "log_mxad"),
  par0 = par0, lower = lower, upper = upper, control.par = NULL)

Arguments

ves
A VES object

opt.method
A character string specifying the optimization method used to estimate the real resisitivity. Currently the following methods are supported:

L-BFGS-B: Limited-memory modification of the BFGS quasi-Newton method (optim package)

SA: Simulated Annealing (GenSA package)

GA: Genetic Algorithms (GA package)

PSO: Particle Swarm Optimization (pso package)

DE: Differential Evoluation (DEoptim package)

obj.fn
Objective function used in the parameter estimation

par0
A numeric vector with the Initial solution

lower
A numeric vector of length equal to the number of layers with the min values of the parameter space

upper
A numeric vector of length equal to the number of layers with the max values of the parameter space

control.par
A list with the parameters of the optimization algorithm

Value

This function returns a list with the following entries:

par: vector with all parameters

value: value of the objective function

rho: Numeric vector with the real resistivities

thickness: Numeric vector with the layer thicknesses

rel.error: The value of the relative error

See also