calibrate_ilsqp

Function to estimate the layer parameters (real resistivities and thicknesses) from a VES measurements using the ILPSQ (Iterative Least-Squares Procedure) with Singular Value Decomposition proposed by Muiuane and Pedersen (1999).

calibrate_ilsqp(ves, iterations = 100, ireport = 10)

Arguments

ves
A VES object

iterations
An integer specifying the maximum number of iterations

ireport
An integer specifying the report interval

Value

This function returns a list with the following entries:

par: A numeric vector with the values of the layer resistivities and thicknesses

value: The value or the RSS (Residual sum of squares)

rel.error: The value of the relative error (in percentage)

cal.error: A matrix with the RSS and relative error at each iteration

References

Muiuane, E. & Pedersen, L. Automatic 1D interpretation of DC resistivity sounding data. Journal of Applied Geophysics, 42, 1, 35-45, 1999.

See also

Other calibration functions: calibrate_joint_nls, calibrate_nls, calibrate_seq_joint_nls, calibrate_seq_nls, calibrate_seq, calibrate_svd, calibrate, log_mnad_resistivity, log_mxad_resistivity, log_rss_resistivity, mnad_resistivity, mxad_resistivity, relative_error_resitivity, rss_resistivity