plot_diagnostic

Function to create an estimation diagnostic plot composed of:

Scatterplot between the measured and calculated apparement resistivity

Scatterplot between calculated apparent resistivity and the residuals

Scatterplot between the calculated apparent resistivity and the standarized residuals

QQplot of the residuals

plot_diagnostic(x)

Arguments

x
A VES object, The VES must be intepreted.

Value

This function returns a ggplot2 object