transform_direct

Function to transform the apparent resistivity into effective resistivities which are assumed closer to the real resistivities using an exponential equation.

transform_direct(ves, k = 2.3, alpha = 0.15)

Arguments

ves
A ves object.

k
Empirical constant.

alpha
Decay exponent (pseudo-porosity factor).

Value

This function returns a list with the following entries:

depth: Depth

real.res: Real resisitivity

Details

The conversion from apparent resitivities to effective resistivities is achieved using the following expression: $$ \rho_{eff}=k\rho_{app}\exp{(-(1-\alpha))} $$ where:

\(\rho_{eff}\) is the effective resistivity which is assumed to be closer to the true resistivity.

\(\rho_{app}\) is the apparent resistivity measured in the field.

\(k\) is a empirical constant defined to improve the fit between the apparent and effective resistivities. This constant has a value of 2.3.

\(\alpha\) is the exponent in the exponential relationship. Its value is of 0.15.

References

Meju, M. Simple effective resistivity-depth transformations for infield or real-time data processing. Computers and Geosciences, 21(8), 985-992, 1995.

See also